+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 ## 3.0.0
 
-**Features / Breaking Changes**
+**Breaking Changes**
 - Refactored Utility Parsers into separate files for better organization.
 - Removed all static utility parsers from Parser class.
+- Removed Character specific parsers from Parser class.
+- Renamed methods in Parser class to be more descriptive and consistent.
+
+**Enhancements**
 - Improved overall parsing performance
+
+**New Features**
+